+++ Makefile
@@ -1,0 +1,3 @@
+build:
+	yarn run test
+	tsc
