+++ package.json
@@ -1,10 +1,9 @@
 {
   "name": "astroboy.ts",
-  "version": "1.0.3-rc.3",
+  "version": "1.0.3-rc.4",
   "description": "基于[astroboy]的DI版本，TypeScript3.2信仰加成",
   "main": "./dist/index.js",
   "scripts": {
-    "build": "tsc",
     "test": "mocha -r ts-node/register src/**/*.spec.ts",
     "cover": "nyc mocha src/**/*.spec.ts"
   },
